test label

+++ dataextract_module.cc
@@ -372,7 +372,7 @@ void test::dataextract::analyze(art::Event const& e)
                     HitTrackID.push_back(trackInfo[0].trackID);
                     hitid=trackInfo[0].trackID;
                     std::cout<<"chan: "<<simChannelNumber<<" i: "<<hitpt<<" infosize: "<<trackInfo.size()<<std::endl;
-                    for(int j=0;j<trackInfo.size();j++){
+                    for(int j=0;j<(int)trackInfo.size();j++){
                         std::cout<<"j: "<<j<<std::endl;
                         //std::cout<<"numElectrons: "<<trackInfo[i].numElectrons<<std::endl;
                         //std::cout<<"energy: "<<trackInfo[i].energy<<std::endl;
